--- src/main/scala-2.11/markets/orders/LimitOrderLike.scala
@@ -1,5 +1,0 @@
-package markets.orders
-
-
-/** Mixin marking an order as a limit order.  Primarily used for pattern matching purposes. */
-trait LimitOrderLike
