--- src/main/scala-2.11/markets/orders/MarketOrderLike.scala
@@ -1,4 +1,0 @@
-package markets.orders
-
-
-trait MarketOrderLike
