+++ src/main/scala-2.11/markets/orders/filled/FilledOrderLike.scala
@@ -13,21 +13,26 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders
+package markets.orders.filled
 
-import akka.actor.ActorRef
+import markets.MessageLike
 import markets.tradables.Tradable
 
+import akka.actor.ActorRef
 
-case class MarketBidOrder(issuer: ActorRef,
-                          quantity: Long,
-                          timestamp: Long,
-                          tradable: Tradable) extends MarketOrderLike with BidOrderLike {
 
-  val price: Long = Long.MaxValue
+trait FilledOrderLike extends MessageLike {
 
-  def split(newQuantity: Long): MarketBidOrder = {
-    MarketBidOrder(issuer, newQuantity, timestamp: Long, tradable)
-  }
+  def counterParties: (ActorRef, ActorRef)
+
+  def tradable: Tradable
+
+  def price: Long
+
+  def quantity: Long
+
+  require(price > 0, "Price must be strictly positive.")
+
+  require(quantity > 0, "Quantity must be strictly positive.")
 
 }
