+++ src/main/scala-2.11/markets/orders/filled/TotalFilledOrder.scala
@@ -13,28 +13,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders
+package markets.orders.filled
 
-import markets.MessageLike
 import markets.tradables.Tradable
-
-import scala.collection.immutable
 
 import akka.actor.ActorRef
 
 
-trait FilledOrderLike extends MessageLike {
-
-  def counterParties: (ActorRef, ActorRef)
-
-  def tradable: Tradable
-
-  def price: Long
-
-  def quantity: Long
-
-  require(price > 0, "Price must be strictly positive.")
-
-  require(quantity > 0, "Quantity must be strictly positive.")
-
-}
+case class TotalFilledOrder(counterParties: (ActorRef, ActorRef),
+                            price: Long,
+                            quantity: Long,
+                            timestamp: Long,
+                            tradable: Tradable) extends FilledOrderLike
