+++ src/main/scala-2.11/markets/orders/filled/package.scala
@@ -13,9 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders.orderings
-
-import markets.orders.BidOrderLike
+package markets.orders
 
 
-object BidTimeOrdering extends TimeOrdering[BidOrderLike]
+package object filled
