+++ src/main/scala-2.11/markets/orders/market/package.scala
@@ -15,18 +15,5 @@ limitations under the License.
 */
 package markets.orders
 
-import akka.actor.ActorRef
-import markets.tradables.Tradable
 
-
-case class LimitAskOrder(issuer: ActorRef,
-                         price: Long,
-                         quantity: Long,
-                         timestamp: Long,
-                         tradable: Tradable) extends LimitOrderLike with AskOrderLike {
-
-  def split(newQuantity: Long): LimitAskOrder = {
-    LimitAskOrder(issuer, price, newQuantity, timestamp, tradable)
-  }
-
-}
+package object market
