+++ src/main/scala-2.11/markets/orders/orderings/ask/AskPricePriority.scala
@@ -1,4 +1,4 @@
-package markets.orders.orderings
+package markets.orders.orderings.ask
 
 import markets.orders.AskOrderLike
