+++ src/main/scala-2.11/markets/orders/orderings/ask/AskTimeOrdering.scala
@@ -13,9 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders.orderings
+package markets.orders.orderings.ask
 
-import markets.orders.BidOrderLike
+import markets.orders.AskOrderLike
+import markets.orders.orderings.TimeOrdering
 
 
-object BidPriceTimeOrdering extends PriceTimeOrdering[BidOrderLike] with BidPricePriority
+object AskTimeOrdering extends TimeOrdering[AskOrderLike]
