+++ src/main/scala-2.11/markets/orders/orderings/bid/BidPricePriority.scala
@@ -1,4 +1,4 @@
-package markets.orders.orderings
+package markets.orders.orderings.bid
 
 import markets.orders.BidOrderLike
