+++ src/main/scala-2.11/markets/orders/orderings/bid/BidTimeOrdering.scala
@@ -13,21 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orders
+package markets.orders.orderings.bid
 
-import akka.actor.ActorRef
-import markets.tradables.Tradable
+import markets.orders.BidOrderLike
+import markets.orders.orderings.TimeOrdering
 
 
-case class MarketAskOrder(issuer: ActorRef,
-                          quantity: Long,
-                          timestamp: Long,
-                          tradable: Tradable) extends MarketOrderLike with AskOrderLike {
-
-  val price: Long = 0
-
-  def split(newQuantity: Long): MarketAskOrder = {
-    MarketAskOrder(issuer, newQuantity, timestamp, tradable)
-  }
-
-}
+object BidTimeOrdering extends TimeOrdering[BidOrderLike]
